+++ packages/frontend/nuxt.config.ts
@@ -52,7 +52,7 @@ export default defineNuxtConfig({
     optimizeDeps: {
       include: [
         'vuedraggable', 'monaco-editor', '@github/webauthn-json/browser-ponyfill', 'randomcolor', '@elastic/apm-rum-vue', '@he-tree/vue',
-        '@iconify/vue/dist/offline', '@iconify/icons-fluent-emoji/face-with-peeking-eye',
+        '@iconify/vue/dist/offline', '@iconify/icons-fluent-emoji/face-with-peeking-eye', 'fabric',
         // Dependencies of @sysreptor/markdown
         '@codemirror/state', '@codemirror/view', '@codemirror/commands', '@codemirror/lint', '@codemirror/autocomplete', 
         '@codemirror/search', '@codemirror/merge', '@codemirror/language', '@codemirror/lang-vue', '@codemirror/lang-css',